Remove metaplex dependency from myMints
also moved around a data class.
