+++ libs/networkInterface/src/main/java/com/solanamobile/mintyfresh/networkinterface/data/MintedMedia.kt
@@ -1,0 +1,8 @@
+package com.solanamobile.mintyfresh.networkinterface.data
+
+data class MintedMedia(
+    val id: String = "",
+    val mediaUrl: String = "",
+    val name: String = "",
+    val description: String = ""
+)
