+++ ui/mymints/build.gradle.kts
@@ -73,9 +73,8 @@ dependencies {
     implementation("com.github.bumptech.glide:compose:1.0.0-alpha.1")
     implementation("com.solanamobile:mobile-wallet-adapter-clientlib-ktx:1.0.4")
 
-    // SolanaKT & Metaplex
+    // SolanaKT
     implementation("com.github.metaplex-foundation:SolanaKT:2.0.0")
-    implementation("com.github.metaplex-foundation:metaplex-android:1.3.0b3")
 
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
